+++ test/e2e/alertmanager_test.go
@@ -857,6 +857,10 @@ func testAlertmanagerConfigCRD(t *testing.T) {
 						},
 						Key: testingSecretKey,
 					},
+					Headers: []monitoringv1alpha1.KeyValue{
+						{Key: "Subject", Value: "subject"},
+						{Key: "Comment", Value: "comment"},
+					},
 				}},
 				VictorOpsConfigs: []monitoringv1alpha1.VictorOpsConfig{{
 					APIKey: &v1.SecretKeySelector{
@@ -1096,6 +1100,9 @@ receivers:
     to: test@example.com
     auth_password: 1234abc
     auth_secret: 1234abc
+    headers:
+      Comment: comment
+      Subject: subject
   pushover_configs:
   - user_key: 1234abc
     token: 1234abc
@@ -1184,5 +1191,112 @@ receivers:
 	})
 	if err != nil {
 		t.Fatalf("%v: %v", err, lastErr)
+	}
+}
+
+func testAMPreserveUserAddedMetadata(t *testing.T) {
+	t.Parallel()
+
+	ctx := framework.NewTestCtx(t)
+	defer ctx.Cleanup(t)
+	ns := ctx.CreateNamespace(t, framework.KubeClient)
+	ctx.SetupPrometheusRBAC(t, ns, framework.KubeClient)
+
+	name := "test"
+
+	alertManager := framework.MakeBasicAlertmanager(name, 3)
+
+	alertManager, err := framework.CreateAlertmanagerAndWaitUntilReady(ns, alertManager)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	updatedLabels := map[string]string{
+		"user-defined-label": "custom-label-value",
+	}
+	updatedAnnotations := map[string]string{
+		"user-defined-annotation": "custom-annotation-val",
+	}
+
+	svcClient := framework.KubeClient.CoreV1().Services(ns)
+	ssetClient := framework.KubeClient.AppsV1().StatefulSets(ns)
+	secretClient := framework.KubeClient.CoreV1().Secrets(ns)
+
+	resourceConfigs := []struct {
+		name   string
+		get    func() (metav1.Object, error)
+		update func(object metav1.Object) (metav1.Object, error)
+	}{
+		{
+			name: "alertmanager-operated service",
+			get: func() (metav1.Object, error) {
+				return svcClient.Get(context.TODO(), "alertmanager-operated", metav1.GetOptions{})
+			},
+			update: func(object metav1.Object) (metav1.Object, error) {
+				return svcClient.Update(context.TODO(), asService(t, object), metav1.UpdateOptions{})
+			},
+		},
+		{
+			name: "alertmanager stateful set",
+			get: func() (metav1.Object, error) {
+				return ssetClient.Get(context.TODO(), "alertmanager-test", metav1.GetOptions{})
+			},
+			update: func(object metav1.Object) (metav1.Object, error) {
+				return ssetClient.Update(context.TODO(), asStatefulSet(t, object), metav1.UpdateOptions{})
+			},
+		},
+		{
+			name: "alertmanager secret",
+			get: func() (metav1.Object, error) {
+				return secretClient.Get(context.TODO(), "alertmanager-test-generated", metav1.GetOptions{})
+			},
+			update: func(object metav1.Object) (metav1.Object, error) {
+				return secretClient.Update(context.TODO(), asSecret(t, object), metav1.UpdateOptions{})
+			},
+		},
+	}
+
+	for _, rConf := range resourceConfigs {
+		res, err := rConf.get()
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		updateObjectLabels(res, updatedLabels)
+		updateObjectAnnotations(res, updatedAnnotations)
+
+		_, err = rConf.update(res)
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	// Ensure resource reconciles
+	alertManager.Spec.Replicas = proto.Int32(2)
+	alertManager, err = framework.UpdateAlertmanagerAndWaitUntilReady(ns, alertManager)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Assert labels preserved
+	for _, rConf := range resourceConfigs {
+		res, err := rConf.get()
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		labels := res.GetLabels()
+		if !containsValues(labels, updatedLabels) {
+			t.Errorf("%s: labels do not contain updated labels, found: %q, should contain: %q", rConf.name, labels, updatedLabels)
+		}
+
+		annotations := res.GetAnnotations()
+		if !containsValues(annotations, updatedAnnotations) {
+			t.Fatalf("%s: annotations do not contain updated annotations, found: %q, should contain: %q", rConf.name, annotations, updatedAnnotations)
+		}
+	}
+
+	if err := framework.DeleteAlertmanagerAndWaitUntilGone(ns, name); err != nil {
+		t.Fatal(err)
 	}
 }